requirements.txt
appdirs==1.4.4
certifi==2022.12.7
charset-normalizer==3.0.1
idna==3.4
importlib-metadata==6.0.0
pyaes==1.6.1
pyasn1==0.4.8
pyee==8.2.2
pyppeteer==1.0.2
python-dotenv==0.21.0
requests==2.28.2
rsa==4.9
Telethon==1.26.1
tqdm==4.64.1
urllib3==1.26.14
websockets==10.4
zipp==3.11.0
